sensors: Create bool to select what timestamp to use
Older devices may have an issue with rotation freezes up and
requires a reboot to fix. In deep sleep the sensor's timestamp
is far off, depending how long it's in sleep, causing rotation
not to work. onSensorChanged if true it will use
SystemClock.elapsedRealtimeNanos() instead of event.timestamp.
Possibly an update to the custom sensor libs.

Change-Id: I7e753867b06b30372d17fc6b0f865a6a05544090

Conflicts:
	core/res/res/values/config.xml